build: add build process for front end js
