+++ .gitignore
@@ -8,6 +8,7 @@ vendor/
 
 # Dev sourcemaps
 admin/build/*.js.map
+public/build/*.js.map
 
 # Jekyll
 docs/_site
